requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
drf-spectacular
django-cors-headers
